+++ History.markdown
@@ -1,3 +1,9 @@
+## HEAD
+
+### Development Fixes
+
+  * test: use categories in post (#249)
+
 ## 0.12.1 / 2019-03-23
 
   * Release: v0.12.0 (#271)